+++ apps/wsdjs_web/lib/wsdjs_web/views/user_view.ex
@@ -10,4 +10,9 @@ defmodule WsdjsWeb.UserView do
   def name_or_email(%User{email: email, name: nil, djname: nil}), do: email
   def name_or_email(%User{name: name}) when is_binary(name), do: name
   def name_or_email(%User{djname: djname}) when is_binary(djname), do: djname
+
+  def user_tooltip(%User{} = user) do
+    "#{user.email}<br/>member since #{Timex.format!(user.inserted_at, "{Mfull} {YYYY}")}"
+  end
+  def user_tooltip(_), do: ""
 end